Remove `quesma/index`
