+++ quesma/clickhouse/clickhouse.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"quesma/concurrent"
 	"quesma/end_user_errors"
-	"quesma/index"
 	"quesma/logger"
 	"quesma/persistence"
 	"quesma/quesma/config"
@@ -272,7 +271,7 @@ func (lm *LogManager) CheckIfConnectedPaidService(service PaidServiceName) (retu
 }
 
 func (lm *LogManager) FindTable(tableName string) (result *Table) {
-	tableNamePattern := index.TableNamePatternRegexp(tableName)
+	tableNamePattern := util.TableNamePatternRegexp(tableName)
 	lm.tableDiscovery.TableDefinitions().
 		Range(func(name string, table *Table) bool {
 			if tableNamePattern.MatchString(name) {
